Refact: Moved logout button to show on profile loading
